requirements.txt
# Основные зависимости для GUI
# tkinter - встроен в Python, не нужно устанавливать

# Системный трей
pystray>=0.19.0
pillow>=8.0.0

# Дополнительные библиотеки (если понадобятся в будущем)
# requests>=2.25.0

